niri: center modal
Was awkward having the prompt appear all the way to the left,
outside the locus of attention.

Think I should probably use eww or Astal for modals, as hacking this
up using terminals feels pretty janky. It limits the amount of styling
and layout I can do, and is hard to integrate into something that
feels cohesive.

On macOS I would use the GUI widgets provided by Hammerspoon for this,
not kitty terminals.