niri: tweak column sizing and positioning
Trying to make it so that windows in the periphery are always partially
visible.

When a window is maximized, I can't tell if there other windows off to
the sides. And moving a maximized window left or right feels jarring,
as I can't see where the other windows are in relation to it.

I want more context.