nvim: increase padding for LINE:COLUMN numbers
Pad line number to 4 digits, and column number to 3 digits.

So that whatever is to the left of the LINE:COLUMN section in the
statusline won't get nudged left/right as you scroll, even when the
file is thousands of lines long, and/or hundreds of characters wide.

I don't like UI elements shifting around. It's distracting.